requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
jsonschema
requests
tqdm
stem
Pillow
selenium
pytest-playwright